First simulation that worked
Simulation worked but aborted due to T0<0. Reason was that wOutlet worked as an inlet. Fix in next iteration of  the simulation. In the middle region of the tube, the mesh refinement was chosen too low. The throttle was too far in. The region refinement around the throttle was too high for practical simulation.